+++ internal/database/connection.go
@@ -1,0 +1,23 @@
+package database
+
+import (
+	"database/sql"
+	"fmt"
+	"github.com/DataStarETL/Orchestrator/internal/config"
+	_ "github.com/lib/pq"
+)
+
+func CreateDBConnection(config *config.DatabaseConfig) (*sql.DB, error) {
+	psqlInfo := fmt.Sprintf("host=%s port=%s user=%s "+
+		"password=%s dbname=%s sslmode=disable",
+		config.Host, config.Port, config.User, config.Password, config.Dbname)
+	db, err := sql.Open("postgres", psqlInfo)
+	if err != nil {
+		return nil, err
+	}
+	err = db.Ping()
+	if err != nil {
+		return nil, err
+	}
+	return db, nil
+}